refactor(example-data): Move Entity Example Data
Entity Example Data
